# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Los Banos Doctors Hospital HR API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /applicants:
    get:
      operationId: listApplicants
      tags: [applicants]
      parameters:
        - in: query
          name: jobId
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Applicant" }
    post:
      operationId: createApplicant
      tags: [applicants]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApplicantInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Applicant" }
  /applicants/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    get:
      operationId: getApplicant
      tags: [applicants]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Applicant" }
    delete:
      operationId: deleteApplicant
      tags: [applicants]
      responses:
        "204":
          description: Deleted
  /applicants/{id}/ai-score:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    post:
      operationId: scoreApplicantAi
      tags: [applicants]
      summary: Run AI resume match scoring for an applicant
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Applicant" }
...
components:
  schemas:
...
    RequirementAnswer:
      type: object
      required: [label, value]
      properties:
        label: { type: string }
        value:
          oneOf:
            - { type: boolean }
            - { type: number }
    RequirementMatch:
      type: object
      required: [label, kind, value, score, weight]
      properties:
        label: { type: string }
        kind: { type: string }
        value:
          oneOf:
            - { type: boolean }
            - { type: number }
        score: { type: number }
        weight: { type: number }
    AiRequirementMatch:
      type: object
      required: [requirement, met, confidence, evidence]
      properties:
        requirement: { type: string }
        met: { type: boolean }
        confidence: { type: number, description: "0 to 1" }
        evidence: { type: string }
        status: { type: string, enum: [full, partial, none] }
        weightPercent: { type: number, description: "Equal share of 100 for this category" }
        pointsAwarded: { type: number, description: "Points this category contributed to the 0-100 score" }
    ApplicantAiEvaluation:
      type: object
      required: [score, summary, matches, model, evaluatedAt]
      properties:
        score: { type: number, description: "0 to 100" }
        summary: { type: string }
        matches:
          type: array
          items: { $ref: "#/components/schemas/AiRequirementMatch" }
        model: { type: string }
        evaluatedAt: { type: string }
        categoryCount: { type: integer }
        scoringExplanation: { type: string }
    Applicant:
      type: object
      required: [id, jobId, name, email, phone, skills, experience, resume, totalScore, matches, createdAt]
      properties:
        id: { type: integer }
        jobId: { type: integer }
        name: { type: string }
        email: { type: string }
        phone: { type: string }
        skills: { type: string }
        experience: { type: string }
        resume: { type: string }
        totalScore: { type: number }
        matches:
          type: array
          items: { $ref: "#/components/schemas/RequirementMatch" }
        aiScore: { type: number, nullable: true }
        aiEvaluation: { $ref: "#/components/schemas/ApplicantAiEvaluation", nullable: true }
        aiUpdatedAt: { type: string, nullable: true }
        createdAt: { type: string }
    ApplicantInput:
      type: object
      required: [jobId, name, email, phone, skills, experience, resume, answers]
      properties:
        jobId: { type: integer }
        name: { type: string }
        email: { type: string }
        phone: { type: string }
        skills: { type: string }
        experience: { type: string }
        resume: { type: string }
        answers:
          type: array
          items: { $ref: "#/components/schemas/RequirementAnswer" }
...