# lib/api-zod/src/generated/types/aiRequirementMatch.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Los Banos Doctors Hospital HR API
 * OpenAPI spec version: 0.1.0
 */

export interface AiRequirementMatch {
  requirement: string;
  met: boolean;
  /** 0 to 1 */
  confidence: number;
  evidence: string;
  /** full = all points, partial = half, none = 0 */
  status?: "full" | "partial" | "none";
  /** Equal share of 100 for this category */
  weightPercent?: number;
  /** Points this category contributed to the 0-100 score */
  pointsAwarded?: number;
}